feat: add policies authentication

+++ Restaurants.Application/Dishes/Commands/DeleteAllForRestaurant/DeleteAllForRestaurantCommandHandler.cs
@@ -2,8 +2,10 @@
 using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Application.Dishes.Queries.GetAllForRestaurant;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteAllForRestaurant
@@ -12,13 +14,16 @@ namespace Restaurants.Application.Dishes.Commands.DeleteAllForRestaurant
         IRestaurantsRepository restaurantsRepository,
         IDishesRepository dishesRepository,
         ILogger<GetAllForRestaurantQueryHandler> logger,
-        IMapper mapper) : IRequestHandler<DeleteAllForRestaurantCommand>
+        IMapper mapper,
+        IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteAllForRestaurantCommand>
     {
         public async Task Handle(DeleteAllForRestaurantCommand request, CancellationToken cancellationToken)
         {
             logger.LogWarning("Deleting all dishes for restaurant with id: {RestaurantId}", request.RestaurantId);
             var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
             if (restaurant == null) throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+            if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+                throw new ForbidException();
             var dishes = restaurant.Dishes;
             if (dishes != null) await dishesRepository.DeleteForRestaurantAsync(dishes);
         }
